feat(cdk-zone-configurations): move into separate module, polish docs, added README.md
