+++ jest.config.js
@@ -9,5 +9,6 @@ module.exports = {
     '<rootDir>/libs/state',
     '<rootDir>/libs/template',
     '<rootDir>/libs/cdk',
+    '<rootDir>/libs/cdk/zone-configurations',
   ],
 };
