+++ libs/cdk/src/index.ts
@@ -3,5 +3,4 @@ export * from './lib/utils';
 export * from './lib/render-strategies';
 export * from './lib/template-management';
 export * from './lib/zone-less';
-export * from './lib/zone-configuration';
 export * from './lib/cdk-config';
