+++ libs/cdk/zone-configurations/README.md
@@ -1,0 +1,38 @@
+# @rx-angular/cdk/zone-configurations
+
+[![npm](https://img.shields.io/npm/v/%40rx-angular%2Fcdk.svg)](https://www.npmjs.com/package/%40rx-angular%2Fcdk)
+![rx-angular CI](https://github.com/rx-angular/rx-angular/workflows/rx-angular%20CI/badge.svg?branch=master)
+[![Coverage Status](https://raw.githubusercontent.com/rx-angular/rx-angular/github-pages/docs/test-coverage/cdk/jest-coverage-badge.svg)](https://rx-angular.github.io/rx-angular/test-coverage/cdk/lcov-report/index.html)
+
+## A small typed and documented convenience helper to configure `zone.js` patching over zone-flags.
+
+`@rx-angular/cdk/zone-configurations` is designed to help developers configure zone-flage easyly.
+Besides a well documented and typed API it also provides a set of convenience features.
+
+## Key features
+
+- ✅ Assertion if zone-flags are not correctly used
+- ✅ Good documentation and maintenance
+- ✅ Typed methods
+- ✅ IDE autocompletion
+- ✅ IDE inline documentation
+- ✅ Predefined event names
+- ✅ Convenience methods
+
+## Demos:
+
+- [⚡ Zone Configuration Demos](https://stackblitz.com/edit/rx-angular-cdk-zone-configuration-demos)
+
+## Install
+
+```bash
+npm install --save @rx-angular/cdk
+# or
+yarn add @rx-angular/cdk
+```
+
+## Documentation
+
+- [Zone Configurations with ZoneFlags](https://github.com/rx-angular/rx-angular/tree/master/libs/cdk/zone-configurations/docs/zone-flags.md)
+  - [How to set up zone flags](https://github.com/rx-angular/rx-angular/tree/master/libs/cdk/zone-configurations/docs/how-to-set-up-zone-flags.md)
+  - [How to debug zone flags](https://github.com/rx-angular/rx-angular/tree/master/libs/cdk/zone-configurations/docs/how-to-debug-zone-flags.md)
