+++ libs/cdk/zone-configurations/docs/images/angular-zone-change-detection_michael-hladky.png
