+++ libs/cdk/zone-configurations/docs/images/angular-zone-flags_disable-timers-false_michael-hladky.png
