+++ libs/cdk/zone-configurations/docs/images/angular-zone-flags_event-listener-patched-flames_michael-hladky.png
