+++ libs/cdk/zone-configurations/docs/images/angular-zone-flags_event-listener-unpatched-flames_michael-hladky.png
