+++ libs/cdk/zone-configurations/docs/images/angular-zone-flags_log-zone-flags_michael-hladky.png
