+++ libs/cdk/zone-configurations/docs/images/angular-zone-flags_performance-comparison_michael-hladky.png
