+++ libs/cdk/zone-configurations/docs/images/angular-zone-flags_setTimeout-patched-flames_michael-hladky.png
