+++ libs/cdk/zone-configurations/docs/images/angular-zone-flags_setTimeout-patched_michael-hladky.png
