+++ libs/cdk/zone-configurations/docs/images/angular-zonejs_timing-marks_detail_michael-hladky.png
