+++ libs/cdk/zone-configurations/docs/images/angular-zonejs_timing-marks_overview_michael-hladky.png
