+++ libs/cdk/zone-configurations/docs/images/angular_zone-flags_disable-timers-true_michael-hladky.png
