+++ libs/cdk/zone-configurations/docs/images/angular_zone-flags_rx-angular_log.png
