+++ libs/cdk/zone-configurations/docs/images/dev-tools_timeline-event-initiators_michael-hladky.png
