+++ libs/cdk/zone-configurations/ng-package.json
@@ -1,0 +1,8 @@
+{
+  "$schema": "../../../node_modules/ng-packagr/ng-package.schema.json",
+  "dest": "../../../dist/libs/cdk/zone-configurations",
+  "lib": {
+    "entryFile": "src/index.ts",
+    "flatModuleFile": "cdk-zone-configurations"
+  }
+}
