+++ libs/cdk/zone-configurations/tslint.json
@@ -1,0 +1,10 @@
+{
+  "extends": "../../../tslint.json",
+  "rules": {
+    "directive-selector": [true, "attribute", "rxAngular", "camelCase"],
+    "component-selector": [true, "element", "rx-angular", "kebab-case"]
+  },
+  "linterOptions": {
+    "exclude": ["!**/*"]
+  }
+}
