+++ nx.json
@@ -33,6 +33,9 @@
     "cdk": {
       "tags": ["type:lib", "type:cdk"]
     },
+    "cdk-zone-configurations": {
+      "tags": ["type:lib", "type:cdk"]
+    },
     "state": {
       "tags": ["type:lib"]
     },
