+++ tsconfig.base.json
@@ -18,14 +18,9 @@
     "paths": {
       "@test-helpers": ["libs/test-helpers/src/index.ts"],
       "@rx-angular/state": ["libs/state/src/index.ts"],
-      "@rx-angular/template": [
-//        "node_modules/@rx-angular/template",
-        "libs/template/src/index.ts"
-      ],
-      "@rx-angular/cdk": [
-//        "node_modules/@rx-angular/cdk",
-        "libs/cdk/src/index.ts"
-      ]
+      "@rx-angular/template": ["libs/template/src/index.ts"],
+      "@rx-angular/cdk": ["libs/cdk/src/index.ts"],
+      "@rx-angular/cdk/zone-configurations": ["libs/cdk/zone-configurations/src/index.ts"]
     }
   },
   "exclude": ["node_modules", "tmp"]
